feat(license-service): refactor smartsolution client; add auth retries (#5050)
