+++ apps/api/src/app/environments/environment.ts
@@ -133,6 +133,10 @@ const devConfig = {
     apiKey: process.env.PKPASS_API_KEY,
     apiUrl: process.env.PKPASS_API_URL,
     secretKey: process.env.PKPASS_SECRET_KEY,
+    cacheKey: process.env.PKPASS_CACHE_KEY ?? 'smartsolution:apitoken',
+    cacheTokenExpiryDelta:
+      process.env.PKPASS_CACHE_TOKEN_EXPIRY_DELTA ?? '2000',
+    authRetries: process.env.PKPASS_AUTH_RETRIES ?? '1',
   },
   audit: {
     defaultNamespace: '@island.is/api',
@@ -269,6 +273,9 @@ const prodConfig = {
     apiKey: process.env.PKPASS_API_KEY,
     apiUrl: process.env.PKPASS_API_URL,
     secretKey: process.env.PKPASS_SECRET_KEY,
+    cacheKey: process.env.PKPASS_CACHE_KEY,
+    cacheTokenExpiryDelta: process.env.PKPASS_CACHE_TOKEN_EXPIRY_DELTA,
+    authRetries: process.env.PKPASS_AUTH_RETRIES,
   },
   audit: {
     defaultNamespace: '@island.is/api',
