+++ libs/api/domains/license-service/src/lib/licenseService.module.ts
@@ -27,6 +27,9 @@ export interface Config {
     apiKey: string
     apiUrl: string
     secretKey: string
+    cacheKey: string
+    cacheTokenExpiryDelta: string
+    authRetries: string
   }
 }
